backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.3",
    "httpx>=0.28.1",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.0",
    "sentence-transformers>=5.1.2",
    "uvicorn>=0.38.0",
]
